精算連動の機能を追加。
精算取引があるものがリンク経由で消えてしまう不具合に対処したつもり。

git-svn-id: svn+ssh://ko.meadowy.net/home/svn/inside/kozuchi/trunk@471 c938ebee-df10-0410-95fc-be8c4b0ebdd5

+++ app/models/settlement.rb
@@ -13,6 +13,10 @@ class Settlement < ActiveRecord::Base
            :foreign_key => 'result_settlement_id',
            :include => :deal
 
+  belongs_to :submitted_settlement
+  
+  belongs_to :user
+
   attr_accessor :result_date, :result_partner_account_id
 
   def target_sum
@@ -21,12 +25,41 @@ class Settlement < ActiveRecord::Base
     sum
   end
   
+  # 相手に提出済にする
+  def submit
+    # すでに提出済ならエラー
+    raise "すでに#{submitted_settlement.user.login}さんに提出されています。" if self.submitted_settlement
+    
+    target_account = account.connected_accounts.first # 1つであることを想定
+    raise "#{account.name}には連携先がありません。" unless target_account
+
+    submitted = nil
+    Settlement.transaction do
+      submitted = SubmittedSettlement.create!(:user_id => target_account.user_id, :account_id => target_account.id, :name => "#{account.name}の精算")
+      # この精算にひもづいた entry に対応する entryを全部これにひもづける
+      for e in target_entries
+        # 本来あるはずのリンクがない（先方が故意に消したなど）場合は新しく作る
+        e.create_friend_deal unless e.linked_account_entry
+        submitted.target_entries << e.linked_account_entry
+      end
+      # この精算の result_entry のひもづけ
+      raise "異常なデータです。精算取引がありません。" unless self.result_entry
+      result_entry.create_friend_deal unless result_entry.linked_account_entry
+      submitted.result_entry = result_entry.linked_account_entry
+      
+      self.submitted_settlement_id = submitted.id
+      self.save!
+    end
+    
+    submitted
+  end
+  
   protected
   
   def validate
-    errors.add_to_base("精算対象取引が１つもありません")if self.target_entries.empty?
+    errors.add_to_base("対象取引が１件もありません。") if self.target_entries.empty?
   end
-  
+    
   def after_save
     self.target_entries.each{|e| e.save!}
     if !self.result_entry && self.result_partner_account_id && self.result_date
@@ -48,7 +81,7 @@ class Settlement < ActiveRecord::Base
         :user_id => self.user_id,
         :date => self.result_date,
         :summary => self.name,
-        :confirmed => false      
+        :confirmed => true # false にすると、相手方の操作によって消されてしまう。リスク低減のためtrueにする      
       )
       result_deal.save!
       self.result_entry = result_deal.account_entries.detect{|e| e.account_id.to_s == self.account_id.to_s}
@@ -57,7 +90,10 @@ class Settlement < ActiveRecord::Base
   end
   
   def before_destroy
-    self.result_entry.deal.destroy
+    # submit されていたら消せない
+    raise "提出済の精算データは削除できません。" if self.submitted_settlement_id
+
+    self.result_entry.deal.destroy if self.result_entry
   end
   
 end
